requirement.txt
streamlit==1.28.1
llama-index==0.9.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
urllib3==1.26.18